+++ config.yml
@@ -1,4 +1,4 @@
-run: "zne"
+run: "vqe"
 fprefix: "test"
 # Variational quantum eigensolver config
 vqe:
